Add scanner package and related classes
