+++ core-lib/src/main/java/org/idpass/smartscanner/lib/scanner/SmartScannerException.kt
@@ -15,13 +15,9 @@
  *
  *
  */
-package org.idpass.smartscanner.lib.platform
+package org.idpass.smartscanner.lib.scanner
 
-import androidx.camera.core.ImageAnalysis
-import androidx.camera.core.ImageProxy
-
-class SmartScannerAnalyzer(private var analyze: ((ImageProxy) -> Unit)) : ImageAnalysis.Analyzer {
-    override fun analyze(imageProxy: ImageProxy) {
-        analyze.invoke(imageProxy)
-    }
+class SmartScannerException : Exception {
+    constructor() : super()
+    constructor(s: String?) : super(s)
 }
